+++ internal/compose/rename_test.go
@@ -114,6 +114,25 @@ func TestRename_Secrets(t *testing.T) {
 	}
 }
 
+func TestRename_Models(t *testing.T) {
+	composeFileURI := fmt.Sprintf("file:///%v", strings.TrimPrefix(filepath.ToSlash(filepath.Join(os.TempDir(), "compose.yaml")), "/"))
+	u := uri.URI(composeFileURI)
+	for _, tc := range modelReferenceTestCases {
+		t.Run(tc.name, func(t *testing.T) {
+			doc := document.NewComposeDocument(document.NewDocumentManager(), u, 1, []byte(tc.content))
+			edits, err := Rename(doc, &protocol.RenameParams{
+				TextDocumentPositionParams: protocol.TextDocumentPositionParams{
+					TextDocument: protocol.TextDocumentIdentifier{URI: composeFileURI},
+					Position:     protocol.Position{Line: tc.line, Character: tc.character},
+				},
+				NewName: "newName",
+			})
+			require.NoError(t, err)
+			require.Equal(t, tc.renameEdits(composeFileURI), edits)
+		})
+	}
+}
+
 func TestRename_Fragments(t *testing.T) {
 	composeFileURI := fmt.Sprintf("file:///%v", strings.TrimPrefix(filepath.ToSlash(filepath.Join(os.TempDir(), "compose.yaml")), "/"))
 	u := uri.URI(composeFileURI)